+++ .evergreen/build-static-test-tarball.sh
@@ -6,7 +6,7 @@ set -o pipefail
 source ./.evergreen/env.sh
 
 export RUSTFLAGS="-C target-feature=+crt-static"
-cargo test ${BUILD_FEATURES} --target x86_64-unknown-linux-gnu get_exe_name -- --ignored
+cargo test ${BUILD_FEATURES} --target x86_64-unknown-linux-gnu get_exe_name
 TEST_BINARY=$(cat exe_name.txt)
 TEST_TARBALL="/tmp/mongo-rust-driver.tar.gz"
 tar czvf ${TEST_TARBALL} ${TEST_BINARY} ./.evergreen
