+++ src/test.rs
@@ -2,10 +2,11 @@
 #![allow(clippy::cast_possible_wrap)]
 
 #[cfg(feature = "dns-resolver")]
-mod atlas_connectivity;
-mod atlas_planned_maintenance_testing;
-#[cfg(feature = "aws-auth")]
-mod auth_aws;
+#[path = "test/atlas_connectivity.rs"]
+mod atlas_connectivity_skip_ci; // requires Atlas URI environment variables set
+#[path = "test/atlas_planned_maintenance_testing.rs"]
+mod atlas_planned_maintenance_testing_skip_ci; // run from the drivers-atlas-testing project
+mod auth;
 mod bulk_write;
 mod change_stream;
 mod client;
@@ -21,9 +22,9 @@ pub(crate) mod csfle;
 mod cursor;
 mod db;
 mod documentation_examples;
-mod happy_eyeballs;
+#[path = "test/happy_eyeballs.rs"]
+mod happy_eyeballs_skip_ci; // requires happy eyeballs server
 mod lambda_examples;
-mod plain_auth;
 pub(crate) mod spec;
 mod timeseries;
 pub(crate) mod util;
